requirements.txt
Flask==2.0.1
pymongo==3.6.1
numpy==1.18.0
Pillow
tensorflow
matplotlib
glob2
Flask-SocketIO==5.1.0
python-engineio
flask-cors
gevent
eventlet
Flask-PyMongo
pymongo